requirements.txt
anthropic
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-llms-anthropic
chromadb
python-dotenv
sentence-transformers
streamlit
requests
newsapi-python